requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
cycler==0.10.0
decorator==4.1.2
h5py==2.7.0
html5lib==0.9999999
Markdown==2.6.8
matplotlib==2.0.2
networkx==1.11
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
protobuf==3.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
scikit-image==0.13.0
scipy==0.19.1
six==1.10.0
tensorflow-gpu==1.2.1
Werkzeug==0.12.2
